+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
+* Support for Swift 4.2 and Xcode 10.
+
 ## [3.0.0](https://github.com/MLSDev/LoadableViews/releases/tag/3.0.0)
 
 * LoadableView no longer sets 'translatesAutoresizingMasksIntoConstraints' property on container view to false. If you plan to use LoadableView without superview, consider turning this property to false manually, because without a container view it will not be able to compute it's size.